Rename: model Team_Fans to UserTeams
