+++ src/models/index.ts
@@ -5,9 +5,9 @@ import Season from './season';
 import Season_Tickets from './season_tickets';
 import Stadium from './stadium';
 import Team from './team';
-import Team_Fans from './team_fans';
 import Ticket from './ticket';
 import User from './user';
+import UserTeams from './user_teams';
 
 const env = (process.env.NODE_ENV as ENV) || 'development';
 const config = dbConfig[env];
@@ -22,15 +22,15 @@ const sequelize = new Sequelize(
 const db = {
   User: User.initialize(sequelize),
   Team: Team.initialize(sequelize),
-  Team_Fans: Team_Fans.initialize(sequelize),
+  UserTeams: UserTeams.initialize(sequelize),
   Season: Season.initialize(sequelize),
   Stadium: Stadium.initialize(sequelize),
   Ticket: Ticket.initialize(sequelize),
   Season_Tickets: Season_Tickets.initialize(sequelize),
 };
 
-User.belongsToMany(Team, { through: 'Team_Fans' });
-Team.belongsToMany(User, { through: 'Team_Fans', as: 'Fans' });
+User.belongsToMany(Team, { through: 'UserTeams' });
+Team.belongsToMany(User, { through: 'UserTeams' });
 
 Team.hasMany(Stadium);
 Stadium.belongsTo(Team);
